Move validate_application_errors into validators directory
