+++ tests/agent_features/test_notice_error.py
@@ -21,11 +21,11 @@ from testing_support.fixtures import (
     reset_core_stats_engine,
     validate_application_error_event_count,
     validate_application_error_trace_count,
-    validate_application_errors,
     validate_transaction_error_event_count,
     validate_transaction_error_trace_count,
 )
 
+from testing_support.validators.validate_application_errors import validate_application_errors
 from testing_support.validators.validate_transaction_errors import validate_transaction_errors
 from newrelic.api.application import application_instance as application
 from newrelic.api.application import application_settings
