+++ tests/testing_support/fixtures.py
@@ -694,44 +694,44 @@ def capture_transaction_metrics(metrics_list, full_metrics=None):
 #     return _validate_transaction_errors
 
 
-def validate_application_errors(errors=None, required_params=None, forgone_params=None):
-    errors = errors or []
-    required_params = required_params or []
-    forgone_params = forgone_params or []
+# def validate_application_errors(errors=None, required_params=None, forgone_params=None):
+#     errors = errors or []
+#     required_params = required_params or []
+#     forgone_params = forgone_params or []
 
-    @function_wrapper
-    def _validate_application_errors(wrapped, instace, args, kwargs):
+#     @function_wrapper
+#     def _validate_application_errors(wrapped, instace, args, kwargs):
 
-        try:
-            result = wrapped(*args, **kwargs)
-        except:
-            raise
-        else:
+#         try:
+#             result = wrapped(*args, **kwargs)
+#         except:
+#             raise
+#         else:
 
-            stats = core_application_stats_engine()
+#             stats = core_application_stats_engine()
 
-            app_errors = stats.error_data()
+#             app_errors = stats.error_data()
 
-            expected = sorted(errors)
-            captured = sorted([(e.type, e.message) for e in stats.error_data()])
+#             expected = sorted(errors)
+#             captured = sorted([(e.type, e.message) for e in stats.error_data()])
 
-            assert expected == captured, "expected=%r, captured=%r, errors=%r" % (expected, captured, app_errors)
+#             assert expected == captured, "expected=%r, captured=%r, errors=%r" % (expected, captured, app_errors)
 
-            for e in app_errors:
-                for name, value in required_params:
-                    assert name in e.parameters["userAttributes"], "name=%r, params=%r" % (name, e.parameters)
-                    assert e.parameters["userAttributes"][name] == value, "name=%r, value=%r, params=%r" % (
-                        name,
-                        value,
-                        e.parameters,
-                    )
+#             for e in app_errors:
+#                 for name, value in required_params:
+#                     assert name in e.parameters["userAttributes"], "name=%r, params=%r" % (name, e.parameters)
+#                     assert e.parameters["userAttributes"][name] == value, "name=%r, value=%r, params=%r" % (
+#                         name,
+#                         value,
+#                         e.parameters,
+#                     )
 
-                for name, value in forgone_params:
-                    assert name not in e.parameters["userAttributes"], "name=%r, params=%r" % (name, e.parameters)
+#                 for name, value in forgone_params:
+#                     assert name not in e.parameters["userAttributes"], "name=%r, params=%r" % (name, e.parameters)
 
-        return result
+#         return result
 
-    return _validate_application_errors
+#     return _validate_application_errors
 
 
 def validate_custom_parameters(required_params=None, forgone_params=None):
